First commit
Add the main project files:

xmlserializer.py -- Main file, is where the XMLSerializerMixin is
defined.
tests.py -- The tests file.
testdata.py The data to be tested (xml and Python objects).
testpkg (dir) -- Includes a sub-package and a Python file with a class
definition to be tested. It's important for testing importing classes in
nested packages.